Use `object.id` as the handle suffix
`HandleBuilder#build` is updated to accept a `hint` parameter, which it may use
to mint a handle. In the default implementation this is simply appended to the
prefix to generate the handle, but other implementations may use it in other
ways (e.g. as a seed) or ignore it altogether.

Clients *should* pass unique hints to builders; generating the same handle
repeatedly given the same hint is allowed!