Fix HandleAssuranceActor to be a well behaved Hyrax Actor
A Hyrax Actor should manually invoke its `next_actor` unless it intends to
terminate and pop out of the actor stack. This better behaved version of the
`HandleAssuranceActor` does this, and also declines to enqueue handle creation
if something lower in the stack terminates `false`. This means that if
`save` (called by `Hyrax::Actors::BaseActor` near the bottom of the stack)
fails, we won't bother to enqueue a Handle job.

Tests are added for the Actor in an `ActionDispatch::MiddlewareStack` context.