Implement template management
Allows users to create new templates and edit existing templates.

The index view provides a _New Template_ button which takes the user to a
form. For existing templates an _Edit_ button is also provided. Users can change
the name of the template, or the values for a changed object from this form;
submission for both is dispatched to the `TemplatesController#update` action.

Not a lot of time has been spent polishing the form, at this point. It's
probable we will want to separate descriptive terms from administrative terms,
and otherwise sharpen up the UI.