Template Preserve Strategy
Implements the `TemplatePreserveStrategy` and adds a factory method to the
abstract strategy class to dispatch strategy instances.

The `Tufts::Template` class and its factory also get a `#behavior` accessor
which they use to build a strategy on demand during apply.

Finally, the behavior sent to `TemplateApplyJob` by the template application
form is forwarded to the `Template` on build. This causes the correct strategy
to be used when `Template#apply` is called by the job.