Connect TemplateUpdate tasks to a Batch
'TemplateUpdate' objects are now `has_one` Batch as `:batchable`. This allows
the `TemplateUpdate` creation task to redirect to the appropriate batch.

Batches are still partially implemented and some of the behavior in
TemplateUpdate should be refactored to depend on the collaborator
batch. Specifically, both TemplateUpdate batchables and batches know job ids,
which creates duplication. Job queueing should also be moved to the batch.

Closes #157. #158 remains partially complete.