Present Batch job status
Allow batch views to inspect the completion of the entire batch. We consider a
batch complete if all the jobs are complete, "In Progress" if any job is working
or complete, and "Queued" otherwise.

If the jobs have fallen out of the cache (72 hours), they and the batch are
considered "Unavailable".

Closes #160.