This commits restores autocomplete for the Undergrad Thesis contribution form.
This form was using Typeahead, but Hyrax has built in autocomplete functionality that
works with local QA vocabs so that was used instead.

There is a feature test searches for 'geol' and checks to see if the autocomplete returns 'Dept. of Geology'

Hyrax doesn't include the autocomplete functionality in a global scope so I modified our global `Tufts` object
to import it and make it available as `Tufts.autocomplete`. Additional fields will need to be added to that function
if they need autocomplete.