This fixes an error with the `/batches` page. The batches page attempts to load XML files that were uploaded through `MiraXmlImporter`:
```
/opt/epigaea/current/log/production.log:F, [2017-11-12T09:12:46.377890 #10059] FATAL -- : [9efc8b3d-ee77-46e4-bcb5-cfc4eceb6561] ActionView::Template::Error (No such file or directory @ rb_sysopen - /opt/epigaea/releases/20171110154758/tmp/metadata/store/mira_xml.xml):
```

On production those files were remove as part of the deployment process. I added some checking in the importer to make
sure that an XML file exists before checking. This raised another error that caused by the importer using `CarrierWave::SanitizedFile` not `File`.

```
/opt/epigaea/current/log/production.log:F, [2017-11-12T09:34:57.481275 #11755] FATAL -- : [475ad141-3c47-4a14-a385-98170c49bfa1] ActionView::Template::Error (undefined method `exist?' for #<CarrierWave::SanitizedFile:0x00555b6fa75fb0>
```

So I also updated the import specs to use `SanitizedFile`.