This uses code implemented in Avalon for marking a test as optional:
https://github.com/avalonmediasystem/avalon/commit/2a5b21dcb7d40c47a77acf6fb24de089944588b5

This is similar to pending, but allows you to specify that a test can pass or fail in a specific
enviroment. This is useful for intermittent failures on Travis that we are getting
for some feature tests. This marks the batch_edit feature specs as optional when
being run in the Travis enviroment.