This commit completes work on refactoring the display of attributes in the edit form and show views.
The PDF create feature spec now fills out every field on the form and checks
for the correct values after the form has been submitted.

This also restores the transcript UI on the Audio and Video forms and has view specs to check that this UI displays on those works, but not on others.

There was an issue with the new `rights_statement` attribute where the form (which is using the default Hyrax edit partial) returns a singular value, but `Hyrax::BasicMetadata` declares it as mutltivalued. I used the same methods to fix this same issue that happens with `title` to `rights_statement`. In the Hyrax specs `rights_statement` is sometimes a singular value and sometimes multi-valued.