Use .env for Handle configuration
Defaults are left in place for development, and (unfortunately) required values
are left in place for the test. We could refactor to draw the config through a
`HandleConfiguration` singleton, or similar; but this seems fine for now.

This is in support of changing the environment for HANDLE_TARGET_HOSTNAME to
`dl.tufts.edu` in the production environment.