Fix the default Hyrax batch create work forms
This commit fixes the batch create work forms by setting the
`Hyrax::Forms::BatchUploadForm.terms` variable to the terms
that we use in forms via an initializer. The `_form_metadata` partial
has also been tweaked so that it displays properly on the batch
create work pages.

This is tested  with a feature spec that adds a `Pdf` work via the
batch upload form.

I think that it is important to keep the default Hyrax batch new work features working.

Fixes #641