Change batch_user to a real email address
Because we have email notifications turned on, we are automatically
emailing the "batch_user", a special account hyrax expects to exist.
If that email is not set to something that can really receive
messages, we will get bounce notifications and our Amazon email
reputation will suffer.

I also include README instructions for how to change this in the
future, once the system is no longer configured to use
curationexperts.com mail settings.