Avoid re-enequing XmlImport items that have already been enqueued
Items that have been or are set to be processed by other jobs could prevously be
re-uploaded, triggering an additional job for the same item. That job would
become stuck in queue, as it failed repeatedly to create an item that already
exists. This avoids the requeue.

A small performance refactor of `Batch::Item` is also included here, to avoid
unneeded Fedora calls while checking for job ids.

Fixes #680.