Display `date_uploaded` and `date_modified` on show page
This adds `date_uploaded` and `date_modified` to the work
show page underneath the title and description. This adds
CSS to style the display of the dates.

The `GenericWorkPresenter` needed to be updated to display
these attributes. They were returning `nil` before but were
present in the `SolrDocument`.

The PDF feature spec is updated to check for the elements showing
up with the correct values.

Closes #613