Fix misspelling of 'Portal' in locale
This commit changes the misspelling of 'Portal' in the Hyrax English
locale file. Previously, it was spelled 'Potral'. There is also
an addition to the feature spec that creates a PDF to check for this
newly spelled-correctly label. The feature spec was not checking
for this label previously, but only from the value that is selected
in the drop-down: 'nowhere'. Now it checks for the string
'Displays in Portal', which is used as a label  in addition to the
value.

In summary,

- 'Displays in Portal' instead of 'Displays in Potral'
- Additional checking for 'Displays in Portal' in the PDF feature spec

Closes #723