Format tech metadata and place into `<mira_import>`
This commit moves the `<technicalMetadata>` node into
the `<mira_import>` node. This is now formatted correctly
and the exporter XML is pretty-printed. This also
adds checking for individual attributes so that the `<technicalMetadata>`
node doesn't have a lot of self closing tags when there is
no value for an attribute.

Closes #519