Add visibility to export
Visibility is already handled in XML and metadata imports via
`Tufts::ImportRecord`. To handle round trips and changes to visibility on
re-import, we need to include visibility on export.

Also adds a spec to check that the XML builder output is well-formed.

Removing visibility entirely from an exported record and reimporting it will
move the visibility to
`Hydra::AccessControls::AccessRight::VISIBILITY_TEXT_VALUE_PRIVATE`.

Closes #624.