Easily create sample data for batch import testing
I wanted to test batch import against files with more than one
object in them, but I didn't want to waste time creating and
maintaining lots of sample files. This rake task makes an import
file with the given number of entries. This will also be useful
for performance testing of large batches.