Check for required fields
1. Move the XML validation check to the validate! method

2. When there are XML validation errors only report the
first one. A single stray element causes six errors because
it cascades -- really, the user just needs to know the
document is malformed and they should go check it with
an XML parser.

3. Add a method to validate! that checks for required fields.
Right now this includes "dc:title", "tufts:displays_in",
and "model:hasModel". Eventually we might want to put this in
a config somewhere.

4. Explicitly add namespaces of the required fields we're
checking for to the in-memory XML document. The way we
declare XML namespaces varies across some of the test files
we're using, and this ensures the validation works across
that variation.

5. Add required displays_in element to sample import file