Remove access to admin areas to non-admin users
This is a refactor of `routes.rb` to enforce access rules
for non-admin users.

Since non-admin users can notifications on the dashboard, this
updates the specs for the sidebar links to ensure that
non-admin users don't see links to admin functionality in the
dashboard.

This commit also ensures that there is a flash message when
users are redirected from paths that they don't have access
to or try access a non-existent path.

Closes #733