Generate unique filenames for imported metadata files
When uploading a metadata file, we had previously been using the original
filename to identify the uploaded content. The users were frequently uploading
content with the same filenames, and the resulting name collisions caused
overwrites. This generates UUIDs for each uploaded metadata file.

Initially, the UUID token for filename is stored on a model variable to avoid
generating the filename repeatedly. We don't need to store it longer than this,
because the generated filename is stored on the model's `#metadata_file` field
once the upload is complete; `Uploader#original_filename` and `#filename` are
`nil` after reloads.

Closes #682.