Only display the first five errors
Only display the first five errors, but provide the total
number of errors.

When we have an import file with too many errors it causes
a cookie overflow exception. Also, chances are good that
people are uploading an auto-generated file that repeats the
same error over and over, so knowing about every single
error instance isn't helpful anyway.

For example, this tests against an import file with 100
entries, all of which are missing a required field.
Before this commit, attempting to import this file would
consistently crash the application.