Add to collections on import
Introduces a `<tufts:collection>` term that accepts collection ids. Ids passed
into this node will be added to the item on create. If the collection does not
exist, the object import fails in the actor stack with
`ActiveFedora::ObjectNotFoundError`.

There is not currently handling for deleting collections, or validating that
collections exist on object upload. Export is also not handled.

Connected to #703.