Allow setting file types on file_sets within a work via XmlImport
Supports setting filetypes in XML batch imports by passing `type="transcript"`
as an attribute on an XML file.

Callers to the Actor Stack's `#create` method can now pass a filename to each of
the `:transcript`, `:representative`, and `:thumbnail` to force the stack to set
these roles when an item is attached to the work.

Unfortunately, we need to overwrite significant parts of the Hyrax ActorStack to
achieve this goal. The attach files part of the stack is run in a background
job, making it challenging to extend. We create customized versions of a
standard Hyrax actor (`CreateWithFilesAndPassTypesActor`) and a job to handle
the setting (`AttachTypedFilesToWorkJob`). The first allows passthrough of the
required arguments to the job. These arguments must be deleted before the work`s
actor is called to prevent errors on the unknown attributes. The Job overwrite
ensures that the roles are set before the `FileSetActor` (which is, confusingly,
neither an actor, nor in the stack) runs and tries to use its own setting
strategies.

ImportRecord is extended to recognize file roles from the `type` XML attribute,
and the import service passes these roles to the actor stack, triggering the
above setting.

Closes #632.