Configure google analytics
Following guide from https://github.com/samvera/hyrax/wiki/Hyrax-Management-Guide#analytics-and-usage-statistics
Document setup in README.
The google analytics account will need to be updated when the
application is deployed to it's production home.

Note that this does NOT activate the stats in the admin
dashboard. That integration is an incomplete feature.
See https://github.com/samvera/hyrax/wiki/Hyrax-Management-Guide#analytics-and-usage-statistics