Add to `Batch#job_ids` instead of overwriting when new jobs are enqueued
Some kinds of Batchables (especially `XmlImport`) can be updated and
re-enqueued. In this case we need to add jobs to `job_ids`, instead of
overwriting as we did previously. This allows new jobs to be added to existing
ones.

This fixes some intermittently failing tests, which are unskipped here.

Attached to #785.