Fix issue where uploaded files are not faded in
If the files in a batch are not faded in they are not submitted
in the form. This uses the file uploaders stop event which is
triggered when files are done uploading and adds the `in` class
to all of the files that have been uploaded. This will ensure that
they are submitted in the form.

Related to #785