Handle collections on MetadataImport
When importing metadata, ensure that items belong exactly to the collections
specified in the import file.

Handling is added to `ImportRecord#build_object`. This is unit tested, and a
feature spec is added to check that an object changes its membership on import
when the job is run through the UI.

Closes #675.