Added the inverse association for batchable models.
This inverse association was needed in a previous commit for the
XmlImport model, so I'm just making the same change to the other models
that use that polymorphic relationship with the Batch model.

I suspect that this will fix the problem with intermittent failing specs
described in stories #752 and #337, but I can't confirm it since I can't
reproduce those failures locally.