Fix specs that fail intermittently
These specs were marked for intermittent failure on Travis.  I suspect
the problem was the test setup, not a problem with the code.

If you load the `dashboard/works` page with no works in your repo, you
can still click the `check_all` button, but the batch buttons won't
appear on the page.  A situation like that would definitely cause these
specs to fail.

I changed the test setup to clean up old records, then create the 2
expected records.  I added a test to make sure we see exactly those 2
records in the table before we run any of the specs.