Fixed a pending spec
It was failing with a validation error because of the timing of the test
setup.  The spec was set up with a stub to simulate a validation error,
but because of the lazy instantiation of the let(:dt), that validation
stub interfered with FactoryGirl's initial creation of the record.

I moved the test for the old name ahead of the validation stub so that
FactoryGirl can create the record before the stub gets set.