Add all possible attributes to technical metadata export
I compared the listing of attributes that we have in this
file with the possible attributes in:
https://github.com/samvera/hydra-works/blob/d41be82540eca9d3a76301df3dcea5b74c9d5937/lib/hydra/works/characterization/fits_document.rb

This captures all of the them that present in the characterization proxy.

Some are missing, and I'll do some investigation and possibly make
tickets in Hyrax Works if there is missing metadata that should be
present.

Closes #627