Make sure fields get marked as dirty if they change. Story #122
I overrode the getter and setter methods for the ordered fields to keep
track of their order in secondary fields, but that caused a problem.
Even though those fields appeared to be correct in the UI, I noticed
that the `description` and `ordered_descriptions` fields were out of
sync in the fedora console.

The overrides to the getters and setters messed up the logic that
determines whether or not a field is marked dirty.

I wrote a spec to make sure that both the original field and its
secondary ordering field will be marked dirty if the field is updated,
and then moved the `super` call to the top of the setter method, which
solves the problem with the logic for dirtiness.