Use models from `Tufts::Curation`
This is an initial cutover to the `epigaea_models` repository. For now, we are
pinning to a tag rather than a gem release (we still need to discuss naming with
Tufts, and may want to let them cut any RubyGems releases).

The core model behaviors are moved to the gem. Metadata setup is moved to
`Tufts::Curation::Schema`. A single indexer has been extracted from the various
indexers here. The dependencies in the gem have initially been left very
thin--it is not a rails engine, and depends on only fragments of Hyrax.

More behavior can still be extracted.