Fix for adding a work to a collection
This test was failing because when HydraEditor runs the
'build_permitted_params' method, the method call
'multiple?(:member_of_collection_ids)' was returning false when it
should return true. This caused that field to be stripped from
the params that are passed to the actor stack, so they never
reach CollectionsMembershipActor.

The problem is fixed by correcting the inheritance of
Tufts::MetadataService so that it inherits from Hyrax, which
in turn inherits from HydraEditor. Crucially,
Hyrax::FormMetadataService, the step in the inheritance
chain that was previously being skipped,  is the place
where :member_of_collection_ids gets defined as multiple.

Hyrax::FormMetadataService also assumes that :rights_statement
is singular, whereas Tufts expects it to be plural, so we
also need to override that setting.