Completed a pending spec for editing a template.
There was a pending spec for the templates controller to check that the
previously-entered values appear on the template's edit form.  That spec
would have been hard to test in a controller spec because I would have
to reach down too many levels.  (e.g. in order to test the `title`, it
would be something like
`assigns(:form).template.changeset.object.title`.)

So, I decided to test it in a feature spec instead.  I added a test to
check that the `title` field in the template appears correctly on the
form when you edit an existing template.