Fix results for Date Created
This fixes an issue of the advanced search form, where
an incorrect solr field for 'Date Created' is used in the
query generated by the form.

This updates the pdf factory with a `date_created` attribute
and changes the the advanced search feature spec to check
that the correct results are returned for a date query.

Related to #915